Clean up LevelConfig API (Phase 1.5) (#16)
* Clean up LevelConfig: consolidate DEFAULT_FPS, fix types, simplify API

- Make DEFAULT_FPS public in LevelConfig; PacmanGame now references it
  instead of duplicating the constant
- Change penLeavingLimits from double[] to int[] to match the integer
  values it holds and PEN_LEAVING_FOOD_LIMITS in PacmanGame
- Compute frightTotalTime in the LevelConfig constructor, removing the
  mutable initFrightTotalTime() method and making the field final
- Rename getLevels() to getLevelConfig() for accuracy (returns a single
  config, not a collection)


* Revert frightTotalTime change: defer to Phase 4 timing refactor

Keep initFrightTotalTime() and its caller in PacmanGame so that 0.23
stays defined only in EVENT_TIME_TABLE. The frightTotalTime cleanup will
be done together with the timing[] named-constant refactor in Phase 4.


* Revert DEFAULT_FPS consolidation: defer to Phase 4

DEFAULT_FPS belongs semantically in PacmanGame (game loop owner), but
moving it there creates a circular dependency with LevelConfig. Will
resolve with a shared constants class in Phase 4 alongside the timing[]
refactor.


---------

+++ app/src/main/java/jp/or/iidukat/example/pacman/LevelConfig.java
@@ -22,7 +22,7 @@ public class LevelConfig {
     private final int fruitScore;
     private final double[] ghostModeSwitchTimes;
     private final int penForceTime;
-    private final double[] penLeavingLimits;
+    private final int[] penLeavingLimits;
     private final int cutsceneId;
 
     static class Builder {
@@ -43,7 +43,7 @@ public class LevelConfig {
         private int fruitScore;
         private double[] ghostModeSwitchTimes;
         private int penForceTime;
-        private double[] penLeavingLimits;
+        private int[] penLeavingLimits;
         private int cutsceneId;
 
         Builder ghostSpeed(double val) { this.ghostSpeed = val; return this; }
@@ -63,7 +63,7 @@ public class LevelConfig {
         Builder fruitScore(int val) { this.fruitScore = val; return this; }
         Builder ghostModeSwitchTimes(double[] val) { this.ghostModeSwitchTimes = val; return this; }
         Builder penForceTime(int val) { this.penForceTime = val; return this; }
-        Builder penLeavingLimits(double[] val) { this.penLeavingLimits = val; return this; }
+        Builder penLeavingLimits(int[] val) { this.penLeavingLimits = val; return this; }
         Builder cutsceneId(int val) { this.cutsceneId = val; return this; }
 
         LevelConfig build() { return new LevelConfig(this); }
@@ -108,12 +108,13 @@ public class LevelConfig {
     void initFrightTotalTime(int blinkDuration) {
         this.frightTotalTime = this.frightTime + blinkDuration * (this.frightBlinkCount * 2 - 1);
     }
+
     public int getFrightBlinkCount() { return frightBlinkCount; }
     public int getFruit() { return fruit; }
     public int getFruitScore() { return fruitScore; }
     public double[] getGhostModeSwitchTimes() { return ghostModeSwitchTimes; }
     public int getPenForceTime() { return penForceTime; }
-    public double[] getPenLeavingLimits() { return penLeavingLimits; }
+    public int[] getPenLeavingLimits() { return penLeavingLimits; }
     public int getCutsceneId() { return cutsceneId; }
 
     static final LevelConfig[] LEVEL_CONFIGS = {
@@ -127,7 +128,7 @@ public class LevelConfig {
                 .frightTime(6).frightBlinkCount(5)
                 .fruit(1).fruitScore(100)
                 .ghostModeSwitchTimes(new double[] { 7, 20, 7, 20, 5, 20, 5, 1, })
-                .penForceTime(4).penLeavingLimits(new double[] { 0, 0, 30, 60, })
+                .penForceTime(4).penLeavingLimits(new int[] { 0, 0, 30, 60, })
                 .build(),
         new LevelConfig.Builder()
                 .ghostSpeed(0.85f).ghostTunnelSpeed(0.45f)
@@ -138,7 +139,7 @@ public class LevelConfig {
                 .frightTime(5).frightBlinkCount(5)
                 .fruit(2).fruitScore(300)
                 .ghostModeSwitchTimes(new double[] { 7, 20, 7, 20, 5, 1033, 1f / 60, 1, })
-                .penForceTime(4).penLeavingLimits(new double[] { 0, 0, 0, 50, })
+                .penForceTime(4).penLeavingLimits(new int[] { 0, 0, 0, 50, })
                 .cutsceneId(1)
                 .build(),
         new LevelConfig.Builder()
@@ -150,7 +151,7 @@ public class LevelConfig {
                 .frightTime(4).frightBlinkCount(5)
                 .fruit(3).fruitScore(500)
                 .ghostModeSwitchTimes(new double[] { 7, 20, 7, 20, 5, 1033, 1f / 60, 1, })
-                .penForceTime(4).penLeavingLimits(new double[] { 0, 0, 0, 0, })
+                .penForceTime(4).penLeavingLimits(new int[] { 0, 0, 0, 0, })
                 .build(),
         new LevelConfig.Builder()
                 .ghostSpeed(0.85f).ghostTunnelSpeed(0.45f)
@@ -161,7 +162,7 @@ public class LevelConfig {
                 .frightTime(3).frightBlinkCount(5)
                 .fruit(3).fruitScore(500)
                 .ghostModeSwitchTimes(new double[] { 7, 20, 7, 20, 5, 1033, 1f / 60, 1, })
-                .penForceTime(4).penLeavingLimits(new double[] { 0, 0, 0, 0, })
+                .penForceTime(4).penLeavingLimits(new int[] { 0, 0, 0, 0, })
                 .build(),
         new LevelConfig.Builder()
                 .ghostSpeed(0.95f).ghostTunnelSpeed(0.5f)
@@ -172,7 +173,7 @@ public class LevelConfig {
                 .frightTime(2).frightBlinkCount(5)
                 .fruit(4).fruitScore(700)
                 .ghostModeSwitchTimes(new double[] { 5, 20, 5, 20, 5, 1037, 1f / 60, 1, })
-                .penForceTime(3).penLeavingLimits(new double[] { 0, 0, 0, 0, })
+                .penForceTime(3).penLeavingLimits(new int[] { 0, 0, 0, 0, })
                 .cutsceneId(2)
                 .build(),
         new LevelConfig.Builder()
@@ -184,7 +185,7 @@ public class LevelConfig {
                 .frightTime(5).frightBlinkCount(5)
                 .fruit(4).fruitScore(700)
                 .ghostModeSwitchTimes(new double[] { 5, 20, 5, 20, 5, 1037, 1f / 60, 1, })
-                .penForceTime(3).penLeavingLimits(new double[] { 0, 0, 0, 0, })
+                .penForceTime(3).penLeavingLimits(new int[] { 0, 0, 0, 0, })
                 .build(),
         new LevelConfig.Builder()
                 .ghostSpeed(0.95f).ghostTunnelSpeed(0.5f)
@@ -195,7 +196,7 @@ public class LevelConfig {
                 .frightTime(2).frightBlinkCount(5)
                 .fruit(5).fruitScore(1000)
                 .ghostModeSwitchTimes(new double[] { 5, 20, 5, 20, 5, 1037, 1f / 60, 1, })
-                .penForceTime(3).penLeavingLimits(new double[] { 0, 0, 0, 0, })
+                .penForceTime(3).penLeavingLimits(new int[] { 0, 0, 0, 0, })
                 .build(),
         new LevelConfig.Builder()
                 .ghostSpeed(0.95f).ghostTunnelSpeed(0.5f)
@@ -206,7 +207,7 @@ public class LevelConfig {
                 .frightTime(2).frightBlinkCount(5)
                 .fruit(5).fruitScore(1000)
                 .ghostModeSwitchTimes(new double[] { 5, 20, 5, 20, 5, 1037, 1f / 60, 1, })
-                .penForceTime(3).penLeavingLimits(new double[] { 0, 0, 0, 0, })
+                .penForceTime(3).penLeavingLimits(new int[] { 0, 0, 0, 0, })
                 .build(),
         new LevelConfig.Builder()
                 .ghostSpeed(0.95f).ghostTunnelSpeed(0.5f)
@@ -217,7 +218,7 @@ public class LevelConfig {
                 .frightTime(1).frightBlinkCount(3)
                 .fruit(6).fruitScore(2000)
                 .ghostModeSwitchTimes(new double[] { 5, 20, 5, 20, 5, 1037, 1f / 60, 1, })
-                .penForceTime(3).penLeavingLimits(new double[] { 0, 0, 0, 0, })
+                .penForceTime(3).penLeavingLimits(new int[] { 0, 0, 0, 0, })
                 .cutsceneId(3)
                 .build(),
         new LevelConfig.Builder()
@@ -229,7 +230,7 @@ public class LevelConfig {
                 .frightTime(5).frightBlinkCount(5)
                 .fruit(6).fruitScore(2000)
                 .ghostModeSwitchTimes(new double[] { 5, 20, 5, 20, 5, 1037, 1f / 60, 1, })
-                .penForceTime(3).penLeavingLimits(new double[] { 0, 0, 0, 0, })
+                .penForceTime(3).penLeavingLimits(new int[] { 0, 0, 0, 0, })
                 .build(),
         new LevelConfig.Builder()
                 .ghostSpeed(0.95f).ghostTunnelSpeed(0.5f)
@@ -240,7 +241,7 @@ public class LevelConfig {
                 .frightTime(2).frightBlinkCount(5)
                 .fruit(7).fruitScore(3000)
                 .ghostModeSwitchTimes(new double[] { 5, 20, 5, 20, 5, 1037, 1f / 60, 1, })
-                .penForceTime(3).penLeavingLimits(new double[] { 0, 0, 0, 0, })
+                .penForceTime(3).penLeavingLimits(new int[] { 0, 0, 0, 0, })
                 .build(),
         new LevelConfig.Builder()
                 .ghostSpeed(0.95f).ghostTunnelSpeed(0.5f)
@@ -251,7 +252,7 @@ public class LevelConfig {
                 .frightTime(1).frightBlinkCount(3)
                 .fruit(7).fruitScore(3000)
                 .ghostModeSwitchTimes(new double[] { 5, 20, 5, 20, 5, 1037, 1f / 60, 1, })
-                .penForceTime(3).penLeavingLimits(new double[] { 0, 0, 0, 0, })
+                .penForceTime(3).penLeavingLimits(new int[] { 0, 0, 0, 0, })
                 .build(),
         new LevelConfig.Builder()
                 .ghostSpeed(0.95f).ghostTunnelSpeed(0.5f)
@@ -262,7 +263,7 @@ public class LevelConfig {
                 .frightTime(1).frightBlinkCount(3)
                 .fruit(8).fruitScore(5000)
                 .ghostModeSwitchTimes(new double[] { 5, 20, 5, 20, 5, 1037, 1f / 60, 1, })
-                .penForceTime(3).penLeavingLimits(new double[] { 0, 0, 0, 0, })
+                .penForceTime(3).penLeavingLimits(new int[] { 0, 0, 0, 0, })
                 .cutsceneId(3)
                 .build(),
         new LevelConfig.Builder()
@@ -274,7 +275,7 @@ public class LevelConfig {
                 .frightTime(3).frightBlinkCount(5)
                 .fruit(8).fruitScore(5000)
                 .ghostModeSwitchTimes(new double[] { 5, 20, 5, 20, 5, 1037, 1f / 60, 1, })
-                .penForceTime(3).penLeavingLimits(new double[] { 0, 0, 0, 0, })
+                .penForceTime(3).penLeavingLimits(new int[] { 0, 0, 0, 0, })
                 .build(),
         new LevelConfig.Builder()
                 .ghostSpeed(0.95f).ghostTunnelSpeed(0.5f)
@@ -285,7 +286,7 @@ public class LevelConfig {
                 .frightTime(1).frightBlinkCount(3)
                 .fruit(8).fruitScore(5000)
                 .ghostModeSwitchTimes(new double[] { 5, 20, 5, 20, 5, 1037, 1f / 60, 1, })
-                .penForceTime(3).penLeavingLimits(new double[] { 0, 0, 0, 0, })
+                .penForceTime(3).penLeavingLimits(new int[] { 0, 0, 0, 0, })
                 .build(),
         new LevelConfig.Builder()
                 .ghostSpeed(0.95f).ghostTunnelSpeed(0.5f)
@@ -296,7 +297,7 @@ public class LevelConfig {
                 .frightTime(1).frightBlinkCount(3)
                 .fruit(8).fruitScore(5000)
                 .ghostModeSwitchTimes(new double[] { 5, 20, 5, 20, 5, 1037, 1f / 60, 1, })
-                .penForceTime(3).penLeavingLimits(new double[] { 0, 0, 0, 0, })
+                .penForceTime(3).penLeavingLimits(new int[] { 0, 0, 0, 0, })
                 .build(),
         new LevelConfig.Builder()
                 .ghostSpeed(0.95f).ghostTunnelSpeed(0.5f)
@@ -307,7 +308,7 @@ public class LevelConfig {
                 .frightTime(0).frightBlinkCount(0)
                 .fruit(8).fruitScore(5000)
                 .ghostModeSwitchTimes(new double[] { 5, 20, 5, 20, 5, 1037, 1f / 60, 1, })
-                .penForceTime(3).penLeavingLimits(new double[] { 0, 0, 0, 0, })
+                .penForceTime(3).penLeavingLimits(new int[] { 0, 0, 0, 0, })
                 .cutsceneId(3)
                 .build(),
         new LevelConfig.Builder()
@@ -319,7 +320,7 @@ public class LevelConfig {
                 .frightTime(1).frightBlinkCount(3)
                 .fruit(8).fruitScore(5000)
                 .ghostModeSwitchTimes(new double[] { 5, 20, 5, 20, 5, 1037, 1f / 60, 1, })
-                .penForceTime(3).penLeavingLimits(new double[] { 0, 0, 0, 0, })
+                .penForceTime(3).penLeavingLimits(new int[] { 0, 0, 0, 0, })
                 .build(),
         new LevelConfig.Builder()
                 .ghostSpeed(0.95f).ghostTunnelSpeed(0.5f)
@@ -330,7 +331,7 @@ public class LevelConfig {
                 .frightTime(0).frightBlinkCount(0)
                 .fruit(8).fruitScore(5000)
                 .ghostModeSwitchTimes(new double[] { 5, 20, 5, 20, 5, 1037, 1f / 60, 1, })
-                .penForceTime(3).penLeavingLimits(new double[] { 0, 0, 0, 0, })
+                .penForceTime(3).penLeavingLimits(new int[] { 0, 0, 0, 0, })
                 .build(),
         new LevelConfig.Builder()
                 .ghostSpeed(0.95f).ghostTunnelSpeed(0.5f)
@@ -341,7 +342,7 @@ public class LevelConfig {
                 .frightTime(0).frightBlinkCount(0)
                 .fruit(8).fruitScore(5000)
                 .ghostModeSwitchTimes(new double[] { 5, 20, 5, 20, 5, 1037, 1f / 60, 1, })
-                .penForceTime(3).penLeavingLimits(new double[] { 0, 0, 0, 0, })
+                .penForceTime(3).penLeavingLimits(new int[] { 0, 0, 0, 0, })
                 .build(),
         new LevelConfig.Builder()
                 .ghostSpeed(0.95f).ghostTunnelSpeed(0.5f)
@@ -352,7 +353,7 @@ public class LevelConfig {
                 .frightTime(0).frightBlinkCount(0)
                 .fruit(8).fruitScore(5000)
                 .ghostModeSwitchTimes(new double[] { 5, 20, 5, 20, 5, 1037, 1f / 60, 1, })
-                .penForceTime(3).penLeavingLimits(new double[] { 0, 0, 0, 0, })
+                .penForceTime(3).penLeavingLimits(new int[] { 0, 0, 0, 0, })
                 .build(),
     };
 }
